# openapi.yaml
openapi: 3.1.0
info:
  title: Preloop AI API
  description: REST API for Preloop AI issue tracking management
  version: 0.1.0
paths:
...
  /api/v1/organizations:
    post:
      tags:
      - Organizations
      summary: Create Organization
      description: Create a new organization, ensuring it's linked to the current
        user's tracker.
      operationId: create_organization_api_v1_organizations_post
      security:
      - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/OrganizationCreate'
      responses:
        '201':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/OrganizationResponse'
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
    get:
      tags:
      - Organizations
      summary: List Organizations
      description: List organizations accessible to the current user.
      operationId: list_organizations_api_v1_organizations_get
      security:
      - bearerAuth: []
      parameters:
      - name: limit
        in: query
        required: false
        schema:
          type: integer
          maximum: 1000
          minimum: 1
          default: 100
          title: Limit
      - name: offset
        in: query
        required: false
        schema:
          type: integer
          minimum: 0
          default: 0
          title: Offset
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema: {}
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
  /api/v1/organizations/{organization_id}:
    get:
      tags:
      - Organizations
      summary: Get Organization
      description: Get an organization by ID, ensuring user has access.
      operationId: get_organization_api_v1_organizations__organization_id__get
      security:
      - bearerAuth: []
      parameters:
      - name: organization_id
        in: path
        required: true
        schema:
          type: string
          title: Organization Id
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/OrganizationResponse'
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
    put:
      tags:
      - Organizations
      summary: Update Organization
      description: Update an organization, ensuring user has access.
      operationId: update_organization_api_v1_organizations__organization_id__put
      security:
      - bearerAuth: []
      parameters:
      - name: organization_id
        in: path
        required: true
        schema:
          type: string
          title: Organization Id
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/OrganizationUpdate'
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/OrganizationResponse'
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
    delete:
      tags:
      - Organizations
      summary: Delete Organization
      description: Delete an organization, ensuring user has access.
      operationId: delete_organization_api_v1_organizations__organization_id__delete
      security:
      - bearerAuth: []
      parameters:
      - name: organization_id
        in: path
        required: true
        schema:
          type: string
          title: Organization Id
      responses:
        '204':
          description: Successful Response
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
  /api/v1/organizations/by-identifier/{identifier}:
    get:
      tags:
      - Organizations
      summary: Get Organization By Identifier
      description: Get an organization by identifier, ensuring user has access.
      operationId: get_organization_by_identifier_api_v1_organizations_by_identifier__identifier__get
      security:
      - bearerAuth: []
      parameters:
      - name: identifier
        in: path
        required: true
        schema:
          type: string
          title: Identifier
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/OrganizationResponse'
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
...
components:
  schemas:
...
    HTTPValidationError:
      properties:
        detail:
          items:
            $ref: '#/components/schemas/ValidationError'
          type: array
          title: Detail
      type: object
      title: HTTPValidationError
...
    OrganizationCreate:
      properties:
        name:
          type: string
          title: Name
          description: Organization name
        identifier:
          type: string
          title: Identifier
          description: Unique identifier for the organization
        description:
          anyOf:
          - type: string
          - type: 'null'
          title: Description
          description: Organization description
        settings:
          anyOf:
          - additionalProperties: true
            type: object
          - type: 'null'
          title: Settings
          description: Organization-wide settings
      type: object
      required:
      - name
      - identifier
      title: OrganizationCreate
      description: Model for creating a new organization.
...
    OrganizationResponse:
      properties:
        name:
          type: string
          title: Name
          description: Organization name
        identifier:
          type: string
          title: Identifier
          description: Unique identifier for the organization
        description:
          anyOf:
          - type: string
          - type: 'null'
          title: Description
          description: Organization description
        settings:
          anyOf:
          - additionalProperties: true
            type: object
          - type: 'null'
          title: Settings
          description: Organization-wide settings
        id:
          type: string
          title: Id
          description: Organization ID
        tracker_id:
          type: string
          title: Tracker Id
          description: Tracker ID
        is_active:
          type: boolean
          title: Is Active
          description: Whether the organization is active
          default: true
        created_at:
          anyOf:
          - type: string
          - type: 'null'
          title: Created At
          description: Creation timestamp
        updated_at:
          anyOf:
          - type: string
          - type: 'null'
          title: Updated At
          description: Last update timestamp
        meta_data:
          additionalProperties: true
          type: object
          title: Meta Data
          description: Additional metadata
      type: object
      required:
      - name
      - identifier
      - id
      - tracker_id
      title: OrganizationResponse
      description: Response model for organization data.
    OrganizationUpdate:
      properties:
        name:
          anyOf:
          - type: string
          - type: 'null'
          title: Name
          description: New organization name
        description:
          anyOf:
          - type: string
          - type: 'null'
          title: Description
          description: New organization description
        settings:
          anyOf:
          - additionalProperties: true
            type: object
          - type: 'null'
          title: Settings
          description: Updated organization settings
      type: object
      title: OrganizationUpdate
      description: Model for updating an organization.
...
    ValidationError:
      properties:
        loc:
          items:
            anyOf:
            - type: string
            - type: integer
          type: array
          title: Location
        msg:
          type: string
          title: Message
        type:
          type: string
          title: Error Type
      type: object
      required:
      - loc
      - msg
      - type
      title: ValidationError
...
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: JWT token for authentication